+++ examples/storybook/stories/Plugins/HtmlPlugin/HtmlWithDomainStory.js
@@ -13,12 +13,9 @@ const HtmlWithDomainStory = ({ contentState }) => {
         pluginHtml({ exposeButtons: [html, adsense], siteDomain: 'https://www.wix.com' }),
         pluginLinkPreview(),
       ]}
-      contentState={contentState}
+      content={contentState}
     >
-      <RichContentEditor
-        initialState={contentState}
-        iframeSandboxDomain="https://richcontent-sttorybook.filesusr.com"
-      />
+      <RichContentEditor iframeSandboxDomain="https://richcontent-sttorybook.filesusr.com" />
     </RicosEditor>
   );
 };