+++ examples/storybook/stories/TestCase/ButtonsTest.js
@@ -1,12 +1,11 @@
 /* eslint-disable import/no-unresolved */
 import React from 'react';
 import { RicosEditor } from 'ricos-editor';
-import { RichContentEditor } from 'wix-rich-content-editor';
 import { pluginDivider } from 'wix-rich-content-plugin-divider';
 import { pluginLinkButton, pluginActionButton } from 'wix-rich-content-plugin-button';
 import { pluginCodeBlock } from 'wix-rich-content-plugin-code-block';
 import { pluginEmoji } from 'wix-rich-content-plugin-emoji';
-import { pluginGallery, GALLERY_TYPE } from 'wix-rich-content-plugin-gallery';
+import { pluginGallery } from 'wix-rich-content-plugin-gallery';
 import { pluginHashtag } from 'wix-rich-content-plugin-hashtag';
 import { pluginHeadersMarkdown } from 'wix-rich-content-plugin-headers-markdown';
 // import { pluginHtml } from 'wix-rich-content-plugin-html';
@@ -33,7 +32,7 @@ const plugins = [
   pluginActionButton(),
   pluginCodeBlock(),
   pluginDivider(),
-  pluginGallery(),
+  pluginGallery({ scrollingElement: () => window }),
   pluginHashtag(),
   // pluginHtml(),
   pluginImage(),
@@ -53,15 +52,7 @@ const plugins = [
 ];
 
 export default () => {
-  const ricosProps = {
-    plugins,
-    locale: 'en',
-    // biSettings: { consumer: 'RCE Standalone' },
-    // instance: mockInstance,
-  };
-
-  const config = {
-    [GALLERY_TYPE]: { scrollingElement: () => window },
+  const toolbarSettings = {
     getToolbarSettings: ({ pluginButtons, textButtons }) => {
       return [
         {
@@ -140,9 +131,14 @@ export default () => {
       ];
     },
   };
-  return (
-    <RicosEditor {...ricosProps}>
-      <RichContentEditor config={config} />
-    </RicosEditor>
-  );
+
+  const ricosProps = {
+    plugins,
+    locale: 'en',
+    toolbarSettings,
+    // biSettings: { consumer: 'RCE Standalone' },
+    // instance: mockInstance,
+  };
+
+  return <RicosEditor {...ricosProps} />;
 };