+++ packages/ricos-common/web/src/RicosEngine.tsx
@@ -111,6 +111,9 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
       content,
       RicosModal,
       onError,
+      mediaSettings = {},
+      linkSettings = {},
+      linkPanelSettings = {},
     } = this.props;
 
     const { strategyProps, previewContent, rawCss } = this.runStrategies();
@@ -119,13 +122,18 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
       toolbarSettings || {};
 
     const { openModal, closeModal, ariaHiddenId } = modalSettings;
+    const { pauseMedia, disableRightClick } = mediaSettings;
+    const { anchorTarget, relValue } = linkSettings;
 
     // any of ricos props that should be merged into child
     const ricosPropsToMerge: RichContentProps = {
       isMobile,
       textToolbarType:
         !isMobile && (textToolbarContainer || useStaticTextToolbar) ? 'static' : 'inline',
-      config: { getToolbarSettings },
+      config: {
+        getToolbarSettings,
+        uiSettings: { disableRightClick, linkPanel: linkPanelSettings },
+      },
       initialState: previewContent || content,
       placeholder,
       onError,
@@ -133,6 +141,9 @@ export class RicosEngine extends Component<EngineProps, EngineState> {
         openModal,
         closeModal,
       },
+      disabled: pauseMedia,
+      anchorTarget,
+      relValue,
     };
 
     const mergedRCProps = merge(strategyProps, _rcProps, ricosPropsToMerge, children.props);