+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -51,8 +51,23 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
   };
 
   render() {
-    const { children, toolbarSettings, ...props } = this.props;
+    const { children, toolbarSettings, draftEditorSettings = {}, ...props } = this.props;
     const { StaticToolbar } = this.state;
+
+    const supportedDraftEditorSettings = Object.entries(draftEditorSettings).map(
+      ([k, v]) =>
+        [
+          'autoCapitalize',
+          'autoComplete',
+          'autoCorrect',
+          'spellCheck',
+          'stripPastedStyles',
+          'handleBeforeInput',
+          'handlePastedText',
+          'handleReturn',
+          'tabIndex',
+        ].includes(k) && v
+    );
 
     const child: RichContentChild =
       children && shouldRenderChild('RichContentEditor', children) ? (
@@ -71,13 +86,14 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
           RicosModal={RicosModal}
           isViewer={false}
           key={'editor'}
-          {...props}
           toolbarSettings={toolbarSettings}
+          {...props}
         >
           {React.cloneElement(child, {
             onChange: this.onChange(child.props.onChange),
             ref: ref => (this.editor = ref),
             editorKey: 'editor',
+            ...supportedDraftEditorSettings,
           })}
         </RicosEngine>
       </Fragment>